TButtonRenderer (so even ToolbarButton can be rendered by custom callback)
