+++ src/Column/Column.php
@@ -18,17 +18,8 @@ use Ublaboo\DataGrid\Traits;
 
 abstract class Column extends FilterableColumn
 {
+	use Traits\TButtonRenderer;
 	use Traits\TLink;
-
-	/**
-	 * @var array
-	 */
-	protected $replacements = [];
-
-	/**
-	 * @var Renderer|NULL
-	 */
-	protected $renderer;
 
 	/**
 	 * @var string
@@ -136,31 +127,6 @@ abstract class Column extends FilterableColumn
 		}
 
 		return $this->getColumnValue($row);
-	}
-
-
-	/**
-	 * Try to render item with custom renderer
-	 * @param  Row   $row
-	 * @return mixed
-	 */
-	public function useRenderer(Row $row)
-	{
-		$renderer = $this->getRenderer();
-
-		if (!$renderer) {
-			throw new DataGridColumnRendererException;
-		}
-
-		if ($renderer->getConditionCallback()) {
-			if (!call_user_func_array($renderer->getConditionCallback(), [$row->getItem()])) {
-				throw new DataGridColumnRendererException;
-			}
-
-			return call_user_func_array($renderer->getCallback(), [$row->getItem()]);
-		}
-
-		return call_user_func_array($renderer->getCallback(), [$row->getItem()]);
 	}
 
 
@@ -325,96 +291,6 @@ abstract class Column extends FilterableColumn
 	public function getName()
 	{
 		return $this->name;
-	}
-
-
-	/**
-	 * Set column replacements
-	 * @param  array $replacements
-	 * @return Column
-	 */
-	public function setReplacement(array $replacements)
-	{
-		$this->replacements = $replacements;
-
-		return $this;
-	}
-
-
-	/**
-	 * Tell whether columns has replacements
-	 * @return bool
-	 */
-	public function hasReplacements()
-	{
-		return (bool) $this->replacements;
-	}
-
-
-	/**
-	 * Apply replacements
-	 * @param  Row   $row
-	 * @return array
-	 */
-	public function applyReplacements(Row $row)
-	{
-		$value = $row->getValue($this->column);
-
-		if ((is_scalar($value) || $value === null) && isset($this->replacements[$value])) {
-			return [true, $this->replacements[$value]];
-		}
-
-		return [false, null];
-	}
-
-
-	/**
-	 * Set renderer callback and (it may be optional - the condition callback will decide)
-	 * @param callable $renderer
-	 */
-	public function setRenderer($renderer, $condition_callback = null)
-	{
-		if ($this->hasReplacements()) {
-			throw new DataGridException(
-				'Use either Column::setReplacement() or Column::setRenderer, not both.'
-			);
-		}
-
-		if (!is_callable($renderer)) {
-			throw new DataGridException(
-				'Renderer (method Column::setRenderer()) must be callable.'
-			);
-		}
-
-		if ($condition_callback != null&& !is_callable($condition_callback)) {
-			throw new DataGridException(
-				'Renderer (method Column::setRenderer()) must be callable.'
-			);
-		}
-
-		$this->renderer = new Renderer($renderer, $condition_callback);
-
-		return $this;
-	}
-
-
-	/**
-	 * Set renderer callback just if condition is truthy
-	 * @param callable $renderer
-	 */
-	public function setRendererOnCondition($renderer, $condition_callback)
-	{
-		return $this->setRenderer($renderer, $condition_callback);
-	}
-
-
-	/**
-	 * Return custom renderer callback
-	 * @return Renderer|null
-	 */
-	public function getRenderer()
-	{
-		return $this->renderer;
 	}
